Amélioration de plusieurs services et interfaces (combos, limites, threads de décompression). Suppression des dépendances inutilisées et nettoyage du code.

+++ UmbraSync/Services/ChatProximityBlendService.cs
@@ -287,9 +287,9 @@ public class ChatProximityBlendService : DisposableMediatorSubscriberBase
     
     private static RawPayload MakeColorPushPayload(Vector4 color)
     {
-        var r = byte.Max((byte)(color.X * 255), (byte)0x01);
-        var g = byte.Max((byte)(color.Y * 255), (byte)0x01);
-        var b = byte.Max((byte)(color.Z * 255), (byte)0x01);
+        var r = byte.Max((byte)(color.X * 255), 0x01);
+        var g = byte.Max((byte)(color.Y * 255), 0x01);
+        var b = byte.Max((byte)(color.Z * 255), 0x01);
         return new RawPayload(unchecked([0x02, 0x13, 0x05, 0xF6, r, g, b, 0x03]));
     }
